requirements.txt
# Local RAG System Dependencies

# Core Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Document Processing
pypdf>=3.17.0
unstructured>=0.10.0

# Embeddings
sentence-transformers>=2.2.0

# Vector Store
pinecone-client>=3.0.0

# LLM
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
